fix(pdp): handle Curio's "proofsetCreated" casing
This is a bug that was introduced in https://github.com/FilOzone/synapse-sdk/pull/111
where we changed "proofset" to "proofSet" everywhere but Curio is specific in
what it sends and we just overwrote that. Ooops. We need to handle
"proofsetCreated" but I'm also allowing for "proofSetCreated" here to give us
room to fix that.

This may become irrelevant with https://github.com/FilOzone/synapse-sdk/pull/129