fix!: revert golden path upload() API + minor fixes
This is a temporary revert of the top-level synapse.storage.upload() API to
default to a single provider and return the original Promise<UploadResult>
type, restoring the original functionality as presented in our docs.
Multi-provider uploads both makes this API too complex to use (mainly thanks to
the return type of allSettled) yet also doesn't present enough information to
make it useful for an API (e.g. callbacks are not individuated).

Ref: https://github.com/FilOzone/synapse-sdk/issues/437